+++ PKHeX.Core/Legality/Tables7.cs
@@ -535,7 +535,7 @@ namespace PKHeX.Core
             // No HMs
         };
 
-        internal static readonly int[] MovePP_SM =
+        internal static readonly byte[] MovePP_SM =
         {
             00,
             35, 25, 10, 15, 20, 20, 15, 15, 15, 35, 30, 05, 10, 20, 30, 35, 35, 20, 15, 20, 20, 25, 20, 30, 05, 10, 15, 15, 15, 25, 20, 05, 35, 15, 20, 20, 10, 15, 30, 35, 20, 20, 30, 25, 40, 20, 15, 20, 20, 20,
